Get Rollbar instance
This gets a rollbar instance that is compatible with the
particular environment, client or Fastboot.

The Rollbar client acts like a single library, but it really is
two libraries mushed into one: one that works in Node and one that
works in the Browser. You can't conflate the two. In particular,
the Browser library can't work in Node, and the Node library can't work
in the browser. In addition, they both expect different access token
types.